refactor(config): modularize configuration system into separate modules
- Split monolithic config.rs into organized modules:
  - config/loader.rs: Configuration loading and management
  - config/question.rs: Question definition and rendering logic
  - config/types.rs: Basic types and enums
  - config/tests.rs: Test cases for config functionality
  - config/mod.rs: Module organization and re-exports
- Improve code organization and maintainability
- Preserve all existing functionality and tests
- Enable better separation of concerns within config system
